requirements.txt
flask
python-dotenv
line-bot-sdk
google-genai
notion-client